Feat: Add resolution for component paths without extensions
#changelog

Added resolution for import paths to components that omit the file
extension.

Wick will attempt to resolve a path like `./path/to/component` to one of
the following:
`./path/to/component.md`
`./path/to/component.wick`
`./path/to/component/index.md`
`./path/to/component/index.wick`

If a path cannot be resolved like above, then an error will be thrown.