Make changes to element index attribute `w:u`.
Elements of composite components now have the value of their `w:u`
index set to the value it is within the outer most component tree.
An index of zero is implied for any element that is root element of
a component, so this change allows for "true" index ordering according
to the scope of outer components.